scripts/requirements.txt
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
tqdm==4.66.1
bcrypt==4.1.2
